moved file into thread
